+++ utils/CMakeLists.txt
@@ -53,6 +53,7 @@ target_sources(utils
     labels.cc
     aws_sigv4.cc
     stream_compressor.cc
+    rest/client.cc
     s3/aws_error.cc
     s3/client.cc
     s3/retryable_http_client.cc
